Expose MySQL client. (#61)
ownCloud doesn't handle migrating an install, which means in order to do
so one must directly access the database. The current snap doesn't
expose a MySQL client, so such a thing is currently difficult. This
commit exposes the MySQL client as an app, which will use the root MySQL
user.

This fixes #37.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>