Upgrade to ownCloud 9.0.1. (#62)
Also started using the .htaccess file shipped in the root of ownCloud.
Note that this was done with a direct include rather than enabling
overrides due to the performance penalty. Typically this would be
unsafe, but the dangers are negated by the fact that the .htaccess file
is read-only.

This fixes #54.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>